backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy
opencv-python-headless
scikit-learn
torch
torchvision
Pillow
scipy
matplotlib